Move fontique/attributes.rs to a shared styled_text.
