+++ fontique/src/attributes.rs
@@ -8,6 +8,7 @@
 use core_maths::CoreFloat;
 
 use core::fmt;
+use styled_text::{Stretch, Style, Weight};
 
 /// Primary attributes for font matching: [`Stretch`], [`Style`] and [`Weight`].
 ///
@@ -40,376 +41,5 @@ impl fmt::Display for Attributes {
             "stretch: {}, style: {}, weight: {}",
             self.stretch, self.style, self.weight
         )
-    }
-}
-
-/// Visual width of a font-- a relative change from the normal aspect
-/// ratio, typically in the range `0.5` to `2.0`.
-///
-/// The default value is [`Stretch::NORMAL`] or `1.0`.
-///
-/// In variable fonts, this can be controlled with the `wdth` [axis].
-///
-/// See <https://fonts.google.com/knowledge/glossary/width>
-///
-/// In CSS, this corresponds to the [`font-width`] property.
-///
-/// [axis]: crate::AxisInfo
-/// [`font-width`]: https://www.w3.org/TR/css-fonts-4/#font-width-prop
-#[derive(Copy, Clone, PartialEq, PartialOrd, Debug)]
-pub struct Stretch(f32);
-
-impl Stretch {
-    /// Width that is 50% of normal.
-    pub const ULTRA_CONDENSED: Self = Self(0.5);
-
-    /// Width that is 62.5% of normal.
-    pub const EXTRA_CONDENSED: Self = Self(0.625);
-
-    /// Width that is 75% of normal.
-    pub const CONDENSED: Self = Self(0.75);
-
-    /// Width that is 87.5% of normal.
-    pub const SEMI_CONDENSED: Self = Self(0.875);
-
-    /// Width that is 100% of normal. This is the default value.
-    pub const NORMAL: Self = Self(1.0);
-
-    /// Width that is 112.5% of normal.
-    pub const SEMI_EXPANDED: Self = Self(1.125);
-
-    /// Width that is 125% of normal.
-    pub const EXPANDED: Self = Self(1.25);
-
-    /// Width that is 150% of normal.
-    pub const EXTRA_EXPANDED: Self = Self(1.5);
-
-    /// Width that is 200% of normal.
-    pub const ULTRA_EXPANDED: Self = Self(2.0);
-}
-
-impl Stretch {
-    /// Creates a new stretch attribute with the given ratio.
-    ///
-    /// This can also be created [from a percentage](Self::from_percentage).
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use fontique::Stretch;
-    /// assert_eq!(Stretch::from_ratio(1.5), Stretch::EXTRA_EXPANDED);
-    /// ```
-    pub fn from_ratio(ratio: f32) -> Self {
-        Self(ratio)
-    }
-
-    /// Creates a stretch attribute from a percentage.
-    ///
-    /// This can also be created [from a ratio](Self::from_ratio).
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use fontique::Stretch;
-    /// assert_eq!(Stretch::from_percentage(87.5), Stretch::SEMI_CONDENSED);
-    /// ```
-    pub fn from_percentage(percentage: f32) -> Self {
-        Self(percentage / 100.0)
-    }
-
-    /// Returns the stretch attribute as a ratio.
-    ///
-    /// This is a linear scaling factor with `1.0` being "normal" width.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use fontique::Stretch;
-    /// assert_eq!(Stretch::NORMAL.ratio(), 1.0);
-    /// ```
-    pub fn ratio(self) -> f32 {
-        self.0
-    }
-
-    /// Returns the stretch attribute as a percentage value.
-    ///
-    /// This is generally the value associated with the `wdth` axis.
-    pub fn percentage(self) -> f32 {
-        self.0 * 100.0
-    }
-
-    /// Returns `true` if the stretch is [normal].
-    ///
-    /// [normal]: Stretch::NORMAL
-    pub fn is_normal(self) -> bool {
-        self == Self::NORMAL
-    }
-
-    /// Returns `true` if the stretch is condensed (less than [normal]).
-    ///
-    /// [normal]: Stretch::NORMAL
-    pub fn is_condensed(self) -> bool {
-        self < Self::NORMAL
-    }
-
-    /// Returns `true` if the stretch is expanded (greater than [normal]).
-    ///
-    /// [normal]: Stretch::NORMAL
-    pub fn is_expanded(self) -> bool {
-        self > Self::NORMAL
-    }
-
-    /// Parses the stretch from a CSS style keyword or a percentage value.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use fontique::Stretch;
-    /// assert_eq!(Stretch::parse("semi-condensed"), Some(Stretch::SEMI_CONDENSED));
-    /// assert_eq!(Stretch::parse("80%"), Some(Stretch::from_percentage(80.0)));
-    /// assert_eq!(Stretch::parse("wideload"), None);
-    /// ```
-    pub fn parse(s: &str) -> Option<Self> {
-        let s = s.trim();
-        Some(match s {
-            "ultra-condensed" => Self::ULTRA_CONDENSED,
-            "extra-condensed" => Self::EXTRA_CONDENSED,
-            "condensed" => Self::CONDENSED,
-            "semi-condensed" => Self::SEMI_CONDENSED,
-            "normal" => Self::NORMAL,
-            "semi-expanded" => Self::SEMI_EXPANDED,
-            "extra-expanded" => Self::EXTRA_EXPANDED,
-            "ultra-expanded" => Self::ULTRA_EXPANDED,
-            _ => {
-                if s.ends_with('%') {
-                    let p = s.get(..s.len() - 1)?.parse::<f32>().ok()?;
-                    return Some(Self::from_percentage(p));
-                }
-                return None;
-            }
-        })
-    }
-}
-
-impl fmt::Display for Stretch {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let value = self.0 * 1000.0;
-        if value.fract() == 0.0 {
-            let keyword = match value as i32 {
-                500 => "ultra-condensed",
-                625 => "extra-condensed",
-                750 => "condensed",
-                875 => "semi-condensed",
-                1000 => "normal",
-                1125 => "semi-expanded",
-                1250 => "expanded",
-                1500 => "extra-expanded",
-                2000 => "ultra-expanded",
-                _ => {
-                    return write!(f, "{}%", self.percentage());
-                }
-            };
-            write!(f, "{}", keyword)
-        } else {
-            write!(f, "{}%", self.percentage())
-        }
-    }
-}
-
-impl Default for Stretch {
-    fn default() -> Self {
-        Self::NORMAL
-    }
-}
-
-/// Visual weight class of a font, typically on a scale from 1.0 to 1000.0.
-///
-/// The default value is [`Weight::NORMAL`] or `400.0`.
-///
-/// In variable fonts, this can be controlled with the `wght` [axis].
-///
-/// See <https://fonts.google.com/knowledge/glossary/weight>
-///
-/// In CSS, this corresponds to the [`font-weight`] property.
-///
-/// [axis]: crate::AxisInfo
-/// [`font-weight`]: https://www.w3.org/TR/css-fonts-4/#font-weight-prop
-#[derive(Copy, Clone, PartialEq, PartialOrd, Debug)]
-pub struct Weight(f32);
-
-impl Weight {
-    /// Weight value of 100.
-    pub const THIN: Self = Self(100.0);
-
-    /// Weight value of 200.
-    pub const EXTRA_LIGHT: Self = Self(200.0);
-
-    /// Weight value of 300.
-    pub const LIGHT: Self = Self(300.0);
-
-    /// Weight value of 350.
-    pub const SEMI_LIGHT: Self = Self(350.0);
-
-    /// Weight value of 400. This is the default value.
-    pub const NORMAL: Self = Self(400.0);
-
-    /// Weight value of 500.
-    pub const MEDIUM: Self = Self(500.0);
-
-    /// Weight value of 600.
-    pub const SEMI_BOLD: Self = Self(600.0);
-
-    /// Weight value of 700.
-    pub const BOLD: Self = Self(700.0);
-
-    /// Weight value of 800.
-    pub const EXTRA_BOLD: Self = Self(800.0);
-
-    /// Weight value of 900.
-    pub const BLACK: Self = Self(900.0);
-
-    /// Weight value of 950.
-    pub const EXTRA_BLACK: Self = Self(950.0);
-}
-
-impl Weight {
-    /// Creates a new weight attribute with the given value.
-    pub fn new(weight: f32) -> Self {
-        Self(weight)
-    }
-
-    /// Returns the underlying weight value.
-    pub fn value(self) -> f32 {
-        self.0
-    }
-
-    /// Parses a CSS style font weight attribute.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use fontique::Weight;
-    /// assert_eq!(Weight::parse("normal"), Some(Weight::NORMAL));
-    /// assert_eq!(Weight::parse("bold"), Some(Weight::BOLD));
-    /// assert_eq!(Weight::parse("850"), Some(Weight::new(850.0)));
-    /// assert_eq!(Weight::parse("invalid"), None);
-    /// ```
-    pub fn parse(s: &str) -> Option<Self> {
-        let s = s.trim();
-        Some(match s {
-            "normal" => Self::NORMAL,
-            "bold" => Self::BOLD,
-            _ => Self(s.parse::<f32>().ok()?),
-        })
-    }
-}
-
-impl Default for Weight {
-    fn default() -> Self {
-        Self::NORMAL
-    }
-}
-
-impl fmt::Display for Weight {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let value = self.0;
-        if value.fract() == 0.0 {
-            let keyword = match value as i32 {
-                100 => "thin",
-                200 => "extra-light",
-                300 => "light",
-                400 => "normal",
-                500 => "medium",
-                600 => "semi-bold",
-                700 => "bold",
-                800 => "extra-bold",
-                900 => "black",
-                _ => return write!(f, "{}", self.0),
-            };
-            write!(f, "{}", keyword)
-        } else {
-            write!(f, "{}", self.0)
-        }
-    }
-}
-
-/// Visual style or 'slope' of a font.
-///
-/// The default value is [`Style::Normal`].
-///
-/// In variable fonts, this can be controlled with the `ital`
-/// and `slnt` [axes] for italic and oblique styles, respectively.
-///
-/// See <https://fonts.google.com/knowledge/glossary/style>
-///
-/// In CSS, this corresponds to the [`font-style`] property.
-///
-/// [axes]: crate::AxisInfo
-/// [`font-style`]: https://www.w3.org/TR/css-fonts-4/#font-style-prop
-#[derive(Copy, Clone, PartialEq, Default, Debug)]
-pub enum Style {
-    /// An upright or "roman" style.
-    #[default]
-    Normal,
-    /// Generally a slanted style, originally based on semi-cursive forms.
-    /// This often has a different structure from the normal style.
-    Italic,
-    /// Oblique (or slanted) style with an optional angle in degrees,
-    /// counter-clockwise from the vertical.
-    Oblique(Option<f32>),
-}
-
-impl Style {
-    /// Parses a font style from a CSS value.
-    pub fn parse(mut s: &str) -> Option<Self> {
-        s = s.trim();
-        Some(match s {
-            "normal" => Self::Normal,
-            "italic" => Self::Italic,
-            "oblique" => Self::Oblique(Some(14.)),
-            _ => {
-                if s.starts_with("oblique ") {
-                    s = s.get(8..)?;
-                    if s.ends_with("deg") {
-                        s = s.get(..s.len() - 3)?;
-                        if let Ok(degrees) = s.trim().parse::<f32>() {
-                            return Some(Self::Oblique(Some(degrees)));
-                        }
-                    } else if s.ends_with("grad") {
-                        s = s.get(..s.len() - 4)?;
-                        if let Ok(gradians) = s.trim().parse::<f32>() {
-                            return Some(Self::Oblique(Some(gradians / 400.0 * 360.0)));
-                        }
-                    } else if s.ends_with("rad") {
-                        s = s.get(..s.len() - 3)?;
-                        if let Ok(radians) = s.trim().parse::<f32>() {
-                            return Some(Self::Oblique(Some(radians.to_degrees())));
-                        }
-                    } else if s.ends_with("turn") {
-                        s = s.get(..s.len() - 4)?;
-                        if let Ok(turns) = s.trim().parse::<f32>() {
-                            return Some(Self::Oblique(Some(turns * 360.0)));
-                        }
-                    }
-                    return Some(Self::Oblique(None));
-                }
-                return None;
-            }
-        })
-    }
-}
-
-impl fmt::Display for Style {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let value = match self {
-            Self::Normal => "normal",
-            Self::Italic => "italic",
-            Self::Oblique(None) => "oblique",
-            Self::Oblique(Some(degrees)) if *degrees == 14.0 => "oblique",
-            Self::Oblique(Some(degrees)) => {
-                return write!(f, "oblique({}deg)", degrees);
-            }
-        };
-        write!(f, "{}", value)
     }
 }
