+++ fontique/src/lib.rs
@@ -48,8 +48,9 @@ mod source_cache;
 
 pub use icu_locid::LanguageIdentifier as Language;
 pub use peniko::Blob;
+pub use styled_text::{Stretch, Style, Weight};
 
-pub use attributes::{Attributes, Stretch, Style, Weight};
+pub use attributes::Attributes;
 pub use collection::{Collection, CollectionOptions, Query, QueryFamily, QueryFont, QueryStatus};
 pub use fallback::FallbackKey;
 pub use family::{FamilyId, FamilyInfo};
