+++ parley/Cargo.toml
@@ -18,8 +18,8 @@ workspace = true
 
 [features]
 default = ["system"]
-std = ["fontique/std", "skrifa/std", "peniko/std"]
-libm = ["fontique/libm", "skrifa/libm", "peniko/libm", "dep:core_maths"]
+std = ["fontique/std", "skrifa/std", "peniko/std", "styled_text/std"]
+libm = ["fontique/libm", "skrifa/libm", "peniko/libm", "styled_text/libm", "dep:core_maths"]
 # Enables support for system font backends
 system = ["std", "fontique/system"]
 accesskit = ["dep:accesskit"]
@@ -32,6 +32,7 @@ fontique = { workspace = true }
 core_maths = { version = "0.1.0", optional = true }
 accesskit = { workspace = true, optional = true }
 hashbrown = { workspace = true }
+styled_text = { workspace = true }
 
 [dev-dependencies]
 tiny-skia = "0.11.4"
